+++ service/controller/resource/statuscondition/create.go
@@ -117,7 +117,7 @@ func (r *Resource) EnsureCreated(ctx context.Context, obj interface{}) error {
 		r.logger.Debugf(ctx, "found %d MachineDeployments for tenant cluster", len(mdList.Items))
 	}
 
-	err = r.computeCreateClusterStatusConditions(ctx, cl, uc, nodes, cpList.Items, mdList.Items)
+	err = r.computeClusterStatusConditions(ctx, cl, uc, nodes, cpList.Items, mdList.Items)
 	if err != nil {
 		return microerror.Mask(err)
 	}
@@ -141,7 +141,7 @@ func (r *Resource) EnsureCreated(ctx context.Context, obj interface{}) error {
 	return nil
 }
 
-func (r *Resource) computeCreateClusterStatusConditions(ctx context.Context, cl apiv1alpha2.Cluster, cr infrastructurev1alpha2.CommonClusterObject, nodes []corev1.Node, controlPlanes []infrastructurev1alpha2.G8sControlPlane, machineDeployments []apiv1alpha2.MachineDeployment) error {
+func (r *Resource) computeClusterStatusConditions(ctx context.Context, cl apiv1alpha2.Cluster, cr infrastructurev1alpha2.CommonClusterObject, nodes []corev1.Node, controlPlanes []infrastructurev1alpha2.G8sControlPlane, machineDeployments []apiv1alpha2.MachineDeployment) error {
 	var desiredVersion string
 	var nodesReady bool
 
@@ -158,10 +158,10 @@ func (r *Resource) computeCreateClusterStatusConditions(ctx context.Context, cl 
 		nodesReady = sameMasterCount && sameWorkerCount && sameVersion
 	}
 
-	return r.writeCreateClusterStatusConditions(ctx, cl, cr, nodesReady, desiredVersion)
+	return r.writeClusterStatusConditions(ctx, cl, cr, nodesReady, desiredVersion)
 }
 
-func (r *Resource) writeCreateClusterStatusConditions(ctx context.Context, cl apiv1alpha2.Cluster, cr infrastructurev1alpha2.CommonClusterObject, nodesReady bool, desiredVersion string) error {
+func (r *Resource) writeClusterStatusConditions(ctx context.Context, cl apiv1alpha2.Cluster, cr infrastructurev1alpha2.CommonClusterObject, nodesReady bool, desiredVersion string) error {
 	status := cr.GetCommonClusterStatus()
 	// After initialization the most likely implication is the tenant cluster
 	// being in a creation status. In case no other conditions are given and no